feat: 쇼케이스 검색 API 추가
- 작업내용
  - 검색 API 구현

- 변경사항
  - Query, Command Controller를 하나로 병합
